tests and readme

+++ data.go
@@ -17,9 +17,9 @@ type SymbolsResponse struct {
 }
 
 type ConvertQuery struct {
-	From   string `json:"from"`
-	To     string `json:"to"`
-	Amount int    `json:"amount"`
+	From   string  `json:"from"`
+	To     string  `json:"to"`
+	Amount float64 `json:"amount"`
 }
 
 type ConvertInfo struct {
